前端上传文件

+++ Picture/picture-picture/src/main/java/com/lz/picture/model/domain/PictureInfo.java
@@ -77,7 +77,7 @@ public class PictureInfo implements Serializable {
      * 宽高比例
      */
     @Excel(name = "宽高比例")
-    private Long picScale;
+    private Double picScale;
 
     /**
      * 图片格式
