+++ Picture/picture-picture/src/main/java/com/lz/picture/model/dto/pictureInfo/PictureInfoInsert.java
@@ -61,7 +61,7 @@ public class PictureInfoInsert implements Serializable {
     /**
      * 宽高比例
      */
-    private Long picScale;
+    private Double picScale;
 
     /**
      * 图片格式
