+++ Picture/picture-picture/src/main/java/com/lz/picture/model/dto/pictureInfo/PictureInfoQuery.java
@@ -58,7 +58,7 @@ public class PictureInfoQuery implements Serializable {
     /**
      * 宽高比例
      */
-    private Long picScale;
+    private Double picScale;
 
     /**
      * 图片格式
